requirements.txt
streamlit
PyPDF2
requests
Pillow
